Bytecoin v.2.0.0 release

+++ include/IWallet.h
@@ -111,7 +111,7 @@ struct TransactionParameters {
   std::vector<std::string> sourceAddresses;
   std::vector<WalletOrder> destinations;
   uint64_t fee = 0;
-  uint64_t mixIn = 0;
+  uint16_t mixIn = 0;
   std::string extra;
   uint64_t unlockTimestamp = 0;
   DonationSettings donation;
